+++ CHANGELOG.md
@@ -717,6 +717,7 @@
 ## [unreleased]
 
 - Show actuals grouped by activity against a specific report
+- Show uploaded transactions after an upload
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-59...HEAD
 [release-59]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-58...release-59